requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
django-cors-headers>=4.3.0
openai>=1.3.0
PyPDF2>=3.0.1
pdfplumber>=0.10.3
requests>=2.31.0
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0


